Develop in Master 4.27.2018 (#13)
* Fix 'unexpected unary operater' bug (#8)

* Add support for  mounting  APFS volumes (#10)

Allows user to specificy an APFS volume id, e.g., disk1s1,
to mount with OS X's `diskutil` command. This can be used
for storing records in an encrypted volume.

* Adds support for mounting Apple dmg files (#12)

* Bug/fix merge (#14)

* Fix 'unexpected unary operater' bug (#8) (#9)

* Develop (#11)

* Fix 'unexpected unary operater' bug (#8)

* Add support for  mounting  APFS volumes (#10)

Allows user to specificy an APFS volume id, e.g., disk1s1,
to mount with OS X's `diskutil` command. This can be used
for storing records in an encrypted volume.

* Adds support for mounting Apple dmg files (#12)